% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{soma_jose}
\alias{soma_jose}
\title{Soma Jose}
\usage{
soma_jose(x, y)
}
\arguments{
\item{x}{um numero para somar}

\item{y}{um numero pare somar}
}
\value{
um numero, resultado da soma
}
\description{
Soma Jose
}
\examples{
soma_jose(4, 7)
}
